infra/functions/generate_sas_token/requirements.txt
azure-keyvault-secrets==4.10.0
azure-identity==1.23.0
requests==2.32.4